Conflicts:
	infinipy2/__init__.py
	infinipy2/core/config.py
	infinisdk/core/bindings.py
	infinisdk/core/field.py
	infinisdk/infinibox/volume.py
	tests/test_infinibox/test_components.py

+++ infinisdk/core/translators_and_types.py
@@ -38,4 +38,4 @@ class MillisecondsDatetimeTranslator(ValueTranslator):
         return arrow.get(value / 1000.0)
 
 MillisecondsDatetimeType = TypeInfo(type=arrow.Arrow,
-                                    api_type=float, translator=MillisecondsDatetimeTranslator())
+                                    api_type=int, translator=MillisecondsDatetimeTranslator())
